[lint] add a .isort.cfg file contrib/buildbot
Summary:
On some systems ISort seems to think build.py and server.py are third-party package rather than local modules.
This is clearly a limitation of ISort, because the python import order respects the order of the paths listed in the `sys.path` list  which by default has the current working directory  as first element (https://docs.python.org/3/library/sys.html#sys.path).

The third-party versions of `build` and `server` seem to be installed as indirect Electrum ABC requirements, which makes this problem apparent for `build`, but it was already doing the wrong sorting for `server`.

Test Plan: `arc lint --everything`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14106