[e.cash] Fix button text color
Summary:
Fix the color changing on hover for the button component. A global link hover definition was overriding the
text color on the button on hover. So defining hover colors on the component

Test Plan:
npm run dev
check the white button at the top of the landing page, the text color should stay white on hover. Compared to an older build
where it turns blue
Then check the 'accent' and 'primary' color props on the same button

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14104