[electrum] use correct Electrum ABC version for DMG filename
Summary:
The version used to be computed using the most recent git tag. Now that we are in the monorepo, this yields the Bitcoin ABC version. Fix it by using electrum's setup.py

Use the same method also for the Appimage and the windows build (although they already detected the correct version), and move this version fetching where it is needed instead of forwarding it through an env variable to the dockers.

This deduplicates the version parsing code, which is now only done in `electrum/electrumabc/version.py`

Test Plan:
```
contrib/build-wine/build.sh
contrib/build-linux/appimage/build.sh
```

On MacOS:
```
contrib/osx/make_osx
```

Check the release files use 5.2.5 in their filename.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14145