[secp256k1] field: Use `restrict` consistently in fe_sqrt
Summary:
That is, use it also in the definition and not only the declaration.

I believe this was the intention of commit
be82bd8e0347e090037ff1d30a22a9d614db8c9f (D19870), but it was omitted there.

This is a backport of [[https://github.com/bitcoin-core/secp256k1/pull/1366 | secp256k1#1366]]

Test Plan: `ninja check-secp256k1`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20073