[ecashaddrjs] Return type as lowercase for use with chronik
Summary: Originally returned the values as uppercase to be consistent with return of `cashaddr.decode` without call for `true`. However, this function is only used to make chronik calls. The all uppercase type always needs to be `.toLowerCase()`'d in apps where this function is used. So, should fix it here.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14163