[e.cash] Add dates to blog
Summary:
Add dates to /blog page tiles
The dates come in timestamp format from the api, so need to convert to a more readable version
with a dateconvert function
note the reason for the publish_date vs publishedAt check is because publishedAt is a field generated
by strapi when you publish a post. Afaik there wasnt a way force this value with the legacy timestamps
when I transfered the webflow data. The webflow timestamps were also in a diffent format from what strapi uses

Test Plan:
npm run test to check new tests
preview the site and check the dates look good on the blog page

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14178