[ecash-herald] Only show to/from addresses if sender or receiver are kinda $$$
Summary:
In general, working to tighten up the size of the msgs and make them look nice on mobile.

Now that we are parsing for addresses that are interesting, we should probably only show the address info for those addresses. Otherwise we just print out a bunch of meaningless strings.

Going forward, opportunity for assigning more info to addresses and then rendering if we know something interesting about an address, e.g. if it's an exchange address or not. But this is a good first step to cut back on msg noise.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14194