refactor: Drop unused fclose() from BufferedFile
Summary:
This was only explicitly used in the tests, where it can be replaced by
wrapping the original raw file pointer into a CAutoFile on creation and
then calling CAutoFile::fclose().

Also, it was used in LoadExternalBlockFile(), where it can also be
replaced by the (implicit call to the) CAutoFile destructor after
wrapping the original raw file pointer in a CAutoFile.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/28483 | core#28483]]
https://github.com/bitcoin/bitcoin/pull/28483/commits/fa389d902fbf5fac19fba8c5d0c5e0a25f15ca63

Test Plan: `ninja all check-all bitcoin-fuzzers bench-bitcoin`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19150