[e.cash] Add animation package and reusable settings
Summary:
Need some nice animations on the page. Adding framer motion package which is a popular tool for creating animations
Wrote a script that can return the animation settings object based on specified parameters
Put it in the components directory, but not sure this is best as its not really a component, but didnt really fit in data either. Open to thoughts there
Added the animations to a few places on the homepage so it can be seen in action

Test Plan:
npm run test
preview the site and check the homepage animations

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14242