contrib: add an extract-electrum.sh script to mirror monorepo commits to github
Summary:
This will help keep the ElectrumABC github repo up to date, which is needed because we need to push tags there to host the releases. It is also convenient for troubleshooting and git-blaming to have a repository with an uninterrupted git history.

Note that the process does not remove the files that were removed in the `[electrum] initial monorepo integration` commit (cirrus.yml, pre-commit-config.yaml...). This will have to be done by manually pushing a commit to the github repo, but I don't anticipate it causing any future conflict.

Test Plan:
Run the script, check that it works with no rebase conflicts and includes all commits.

`cd electrumabc-mirror && git log --pretty=oneline`

Compare the electrum directory with the electrumabc-mirror directory with recursive diff or meld. All files not listed in `.gitignore` should be identical (with the exception of the extra files mentionned in the summary that still need deleting in the github repository)

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14253