[alias-server] Deprecate isRegistered: true
Summary:
If you have a blockheight and a txid, `isRegistered: true` is redundant.

Preserving `isRegistered: false` for aliases that are not registered. This is useful information, esp as you can only get it if you query a valid alias.

The meaning of a returned object with just `{alias}` is not intuitively understandable.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14251