[ecash-herald] Add txids to mocks instead of blocks
Summary:
Simplify testing and mocks by building a mock block of interesting txids instead of relying on existing blocks

As the app grew in complexity, actual API data for some blocks became impractical to store (for example, large utxo sets). This data is not necessary to test app functionality and can be mocked / simplified.

We are really interested in parsing on a per-txid basis, not on a per-block basis. With this new script, we can add new interesting txids to prove that the app parses them, without outsized growth in the legacy `blocks.js` mock.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14265