[electrum] add a build configuration for functional tests
Summary:
Run functional / regtest tests to test some Electrum ABC RPC commands and the interaction with Fulcrum and the node.
This can be run occasionnaly to test the latest releases of the node and fulcrum still work well together, and called via buildbot when (rarely) a change is made to an Electrum ABC RPC command.

Test Plan:
`@bot electrum-functional-tests`

On a machine with docker service running:
```
./contrib/teamcity/build-configurations.py electrum-functional-tests
```

Reviewers: #bitcoin_abc, bytesofman, Fabien

Reviewed By: #bitcoin_abc, bytesofman, Fabien

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14275