[electrum] convert the plugin's README.rst to markdown
Summary:
This will fix a	 dead link  on https://www.bitcoinabc.org/electrum/CONTRIBUTING.html and use the same doc format everywhere.

I'm not sure I agree about everything in this document, but let's not mix format changes with documentation updates. The only changes in the document this diff introduces are:
- a couple of `Electron Cash --> Electrum ABC` fixes (undetected previously because they were split over two lines)
- remove the note about the oldest version that supports external plugins (this was before the BCH split, so no such Electrum ABC version exists))

Test Plan: Check rendering in my IDE, test that the links are not dead

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14283