[e.cash] Glitch text updates
Summary:
I've been working on the weglot translation integration and one of the biggest headaches is getting the data-text attribute
to update when the language changes. Weglot cant access attributes like that afaik, so it would require some jankish
js to iterate over all of the glitch text elements and update the data-text element (like we did on the webflow site). The problem is you quickly run into some
react/next lifecycle issues where it doesnt update on the first render but only sometimes and changing pages is whole other
problem.

Long way of saying I propose we move the "glitch" text into the DOM. Gets rid of any headaches with updating the data-text
and should be indistinguishable to the end user. Open to any reasons this is a bad idea though

Also in this glitch text update is a theme value so we can adjust the animation timing between themes

Test Plan: preview the site and check the glitch text behaves as before

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14279