[e.cash] Add Weglot
Summary:
Adding the weglot transaltions
Added the needed scripts and weglot api key (this needs to be client side I believe and was exposed before fyi)
And added some styles to override defualt styles

I went ahead and just used the weglot instance that was setup for the webflow site. It had all of the subdmains setup etc,
so hopefully will just work when deployed ha but really not a great way to test this afaik as its reliant on the e.cash domain...
It only kind of works rn locally, but I believe its because of the subdomain settings, but the language selector does appear
and look how it should
And the subdomains are still working on the live site. Kind of interesting actually, if you go to subdomains rn it flashes
translations, and if you go to the blog they're translated o___o So weglot is tied in there somehow even though the code
isnt on the site...90% sure its coming from the subdomains that are configured in weglot. So all of the blogs got translated
which smoked through our word count in weglot...but I added a rule to ignore /blog going forward so hopefully reverts? unclear
how that works

Test Plan:
preview the site and see if the language selector looks right and possibly functions
Afraid we may have to test live to see whats going on though

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14292