[e.cash] Add social card image
Summary: Adding the social card image back in

Test Plan:
Social cards only work on the prod site and require an absolute path for the image. So not a great way to test.
But can preview the site and check the code is in the head as expected

Once live we can run a validator to see if the image is coming through

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14303