[doc/standards] Improve op return prefix guidelines
Summary:
T3193

Update doc to remove dead link for SLP spec. Remove confusing and incorrect little-endian descriptor for prefixes. Add simple payload descriptions for simple protocols used in Cashtab that probably do not need a full spec.

Test Plan: Review for typos, judgment on whether or not this is an improvement

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14315