[e.cash] Announcement bar edits
Summary:
In preparing the stealth mode switch-over I noticed the announcement bar at the top of the nav could be made more flexible.
Right now it requires it to be a link, but maybe we dont want it to be a link (or dont have a link for it)
It also does not utilize the custom link component so that external links are automatically opened in new tabs.
So modifying it to a component and using custom link and conditional rendering based on the props

Test Plan:
Preview the site, and check out the announcement bar. It should look and function the same
try npm run dev - go to navbar component in an editor, and try changing the announcement bar href prop to /get-ecash.
It should not open in new tab now
try removing the href prop
It should become not a link and have no hover effect

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14336